refactoring checkout details in overview and success into own component
