+++ src/views/CheckoutOverview.vue
@@ -1,15 +1,6 @@
 <template>
     <div class="checkout-overview">
-        <h1>You're sending <Amount :amount="request.value + request.fee"/> to</h1>
-        <Account :address="request.recipient.toUserFriendlyAddress()" :label="originDomain"/>
-        <div v-if="plainData" class="data">{{ plainData }}</div>
-        <div class="sender-section">
-            <div class="sender-nav">
-                <h2>Pay with</h2>
-                <button @click="changeAccount">Change</button>
-            </div>
-            <Account v-if="activeAccount" :address="activeAccount.userFriendlyAddress" :label="activeAccount.label" :balance="activeAccount.balance"/>
-        </div>
+        <CheckoutDetails :accountChangeable="true" /> 
         <div class="page-footer">
             <button @click="proceed">Pay <Amount :amount="request.value + request.fee"/></button>
         </div>
@@ -19,41 +10,22 @@
 <script lang="ts">
 import {Component, Emit, Vue} from 'vue-property-decorator';
 import {Getter} from 'vuex-class';
-import {Amount, Account} from '@nimiq/vue-components';
 import {SignTransactionRequest as KSignTransactionRequest} from '@nimiq/keyguard-client';
 import {State as RpcState} from '@nimiq/rpc';
+import CheckoutDetails from '../components/CheckoutDetails.vue';
 import {AddressInfo} from '../lib/AddressInfo';
 import {KeyInfo} from '../lib/KeyInfo';
 import {RequestType, ParsedCheckoutRequest} from '../lib/RequestTypes';
 import RpcApi from '../lib/RpcApi';
 import staticStore, {Static} from '../lib/StaticStore';
 
-@Component({components: {Amount, Account}})
+@Component({components: {CheckoutDetails}})
 export default class CheckoutOverview extends Vue {
     @Static private rpcState!: RpcState;
     @Static private request!: ParsedCheckoutRequest;
 
     @Getter private activeKey!: KeyInfo | undefined;
     @Getter private activeAccount!: AddressInfo | undefined;
-
-    private get originDomain() {
-        return this.rpcState.origin.split('://')[1];
-    }
-
-    private get plainData() {
-        if (!this.request.data) return undefined;
-
-        try {
-            return Nimiq.BufferUtils.toAscii(this.request.data);
-        } catch (err) {
-            return undefined;
-        }
-    }
-
-    @Emit()
-    private changeAccount() {
-        this.$router.push({name: `${RequestType.CHECKOUT}-change-account`});
-    }
 
     // TODO Figure out how this will be called
     @Emit()
@@ -86,7 +58,7 @@ export default class CheckoutOverview extends Vue {
             recipientLabel: undefined, // TODO: recipient label
             value: this.request.value,
             fee: this.request.fee || 0, // TODO: proper fee estimation
-            validityStartHeight: 176450, // TODO: get valid start height
+            validityStartHeight: 286890, // TODO: get valid start height
             data: this.request.data,
             flags: this.request.flags,
             networkId: this.request.networkId,
@@ -110,61 +82,6 @@ export default class CheckoutOverview extends Vue {
         display: flex;
         flex-direction: column;
         flex-grow: 1;
-    }
-
-    h1 {
-        font-size: 3rem;
-        line-height: 3.625rem;
-        font-weight: 300;
-        letter-spacing: 0.021em;
-        margin: 4rem 4rem 1rem 4rem;
-    }
-
-    h1 .amount {
-        font-weight: 500;
-    }
-
-    .data {
-        font-size: 2rem;
-        line-height: 1.3;
-        opacity: 0.7;
-        padding: 1rem 4rem;
-    }
-
-    .sender-section {
-        margin-top: 3rem;
-        padding-bottom: 2rem;
-        border-top: solid 1px #f0f0f0;
-        border-bottom: solid 1px #f0f0f0;
-        background: #fafafa;
-    }
-
-    .sender-nav {
-        padding: 2rem 2rem 1rem 2rem;
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-    }
-
-    .sender-nav h2 {
-        margin: 1rem;
-        font-size: 1.75rem;
-        text-transform: uppercase;
-        line-height: 0.86;
-        letter-spacing: 0.143em;
-        font-weight: 400;
-    }
-
-    .sender-nav button {
-        background: #e5e5e5;
-        padding: 1rem 1.75rem;
-        width: unset;
-        font-size: 1.75rem;
-        line-height: 0.86;
-        box-shadow: unset;
-        text-transform: unset;
-        letter-spacing: normal;
-        height: auto;
     }
 
     .page-footer {
